+++ TouhouEnemyModels/MyPluginInfo.cs
@@ -4,6 +4,6 @@ namespace TouhouEnemyModels
     {
         public const string PLUGIN_GUID = "com.bean.TouhouEnemyModels";
         public const string PLUGIN_NAME = "TouhouEnemyModels";
-        public const string PLUGIN_VERSION = "0.9.0";
+        public const string PLUGIN_VERSION = "0.9.3";
     }
 }